+++ src/paradigma/Usuario.java
@@ -3,11 +3,11 @@ package paradigma;
 public class Usuario {
 	private String nombre;
 	private int tiempoDisponibleEnHoras;
-	private double presupuesto;
+	private int presupuesto;
 	private TipoDeAtraccion atraccionPreferida;
 
 
-	public Usuario(String nombre, int tiempoDisponibleEnHoras, double presupuesto, TipoDeAtraccion atraccionPreferida) {
+	public Usuario(String nombre, int tiempoDisponibleEnHoras, int presupuesto, TipoDeAtraccion atraccionPreferida) {
 		this.nombre = nombre;
 		this.tiempoDisponibleEnHoras = tiempoDisponibleEnHoras;
 		this.presupuesto = presupuesto;
@@ -20,7 +20,7 @@ public class Usuario {
 	public int getTiempoDisponibleEnHoras() {
 		return this.tiempoDisponibleEnHoras;
 	}
-	public double getPresupuesto() {
+	public int getPresupuesto() {
 		return this.presupuesto;
 	}
 	public TipoDeAtraccion getAtraccionPreferida() {
@@ -39,7 +39,7 @@ public class Usuario {
 		return disponeDePresupuesto && disponeDeTiempo;
 	}
 	
-	public synchronized void restarPresupuestoYtiempo(double precio, int horas) {
+	public synchronized void restarPresupuestoYtiempo(int precio, int horas) {
 		this.presupuesto -= precio;
 		this.tiempoDisponibleEnHoras -= horas;
 	}
